+++ src/Ext/WeaponType/Body.cpp
@@ -59,6 +59,8 @@ void WeaponTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->OmniFire_TurnToTarget.Read(exINI, pSection, "OmniFire.TurnToTarget");
 	this->ExtraWarheads.Read(exINI, pSection, "ExtraWarheads");
 	this->ExtraWarheads_DamageOverrides.Read(exINI, pSection, "ExtraWarheads.DamageOverrides");
+	this->AmbientDamage_Warhead.Read(exINI, pSection, "AmbientDamage.Warhead");
+	this->AmbientDamage_IgnoreTarget.Read(exINI, pSection, "AmbientDamage.IgnoreTarget");
 }
 
 template <typename T>
@@ -84,6 +86,8 @@ void WeaponTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->OmniFire_TurnToTarget)
 		.Process(this->ExtraWarheads)
 		.Process(this->ExtraWarheads_DamageOverrides)
+		.Process(this->AmbientDamage_Warhead)
+		.Process(this->AmbientDamage_IgnoreTarget)
 		;
 };